+++ src/App.tsx
@@ -1,10 +1,10 @@
 import Router from './routes/Router';
 import GlobalStyles from './styles/globalStyles';
-import useRecipeData from './hooks/useRecipeData';
+import { ProcessedRecipeData } from './apis/recipe';
 
-function App() {
+const App = () => {
   // 앱 실행 시 1회 호출하는 레시피 데이터
-  useRecipeData();
+  ProcessedRecipeData();
 
   return (
     <>
@@ -12,6 +12,6 @@ function App() {
       <Router />
     </>
   );
-}
+};
 
 export default App;
