+++ src/components/my/ProfileBox.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import COLORS from '../../styles/colors';
-import { authService } from '../../api/firebase';
+import { authService } from '../../apis/firebase';
 import { User } from 'firebase/auth';
 import useConfirm from '../../hooks/useConfirm';
 import ConfirmModal from '../common/ConfirmModal';
